Update patch set 3
Patch Set 3: (1 inline comment)

> Although it is not ideal, for this fuzzing test purpose -- allowing decoder
> to end gracefully without a crash when an error occurs, the current
> setjmp/longjmp implementation works ok. The only issue I can see now is we
> have to end the threads first, and then free the allocated memory to ensure
> no crash would happen.

Of course, that wasn't my point. This is a good fix, though it'd be nice if this were caught by regular tests rather than the fuzzing. You should add that file to the invalid test case and make sure you can reproduce the failure in the unit tests.

Patch-set: 3