Use normal rate_correction_factor for gf in CBR mode
I0c5f010 changed to allow update golden reference buffer in CBR mode,
this commit changes the use of rate_correction_factor for those frames
to be aligned with the new usage. This commit attempts to solve two
issues:

a. Initialization of rate correction factor for Golden Frame
Prior to this patch, even the regular inter frame has been update
the rate correction factor based on content and encoding results,
the first golden frame would still use the ininitialized value
that can be way off.

b. Allowing the update to be slightly faster
Prior to this patch, when the rate correction factor is way, the
update to the factor is too slow, the factor could not get close
to correct value even after many frames.

The commit helps all clips, but especially to one clips in RTC
set that rate correction was way off: thaloundeskmtgvga with
both overall and average psnr gained 0.5dB.

Change-Id: I0be5c41691be57891d824505348b64be87fa3545