DPCM intra prediction experiment for lossless screen content coding
This patch is built on the non-transform coding experiment. If
transform is skipped, certain intra predictions modes are performed
with pixel-wise DPCM. These modes include: horizontal, vertical, TM,
d135 and d45.

This patch only considers lossless compression. For natrual images,
such as those in the derf dataset, DPCM does not affect performance
much. It is very useful for screen content videos. I tested a total
of 16 screen content clips. The gain ranges from 1.6% ~ 29.2%, and
the average gain is 10.8%.

Change-Id: Ia0b6757dd45640b654db56a403b93c08b1930429