tx_skip mode for lossy coding
This patch improves the non-transform coding mode. At this
point, the coding gain on screen content videos is about
12% for lossless, an 15% for lossy case.

1. Encode tx_skip flags with context. Y tx_skip flag context is
whether the prediction mode is inter or intra. UV flag context
is Y flag.

2. Transform skipping for inter blocks is not helpful unless
the Q-index is extremely high. So in the lossy case, transform
skipping is only enabled for intra blocks. In the lossless case,
it is enabled for both intra and inter blocks.

3. The shift of the prediction residue, when copying them to the
coeff buffer, is set as 3 for the lossy case, and 2 for the
lossless case. This setting yields the best test results.

Change-Id: I372973c7518cf385f6e542b22d0f803016e693b0