Improve tx_skip mode
This patch improves the non-transform coding mode:
1. Fix bugs in the distortion calculation of tx_skip modes.
2. Encode tx_skip flags with context. Y tx_skip flag context is
whether the prediction mode is inter or intra. UV flag context
is Y flag.
3. Transform skipping is more useful when the bitrate is higher.
It is therefore enabled only when the q-index is smaller than a
threshold (currently set as 20). This value may be adjusted,
especially if we create new q-table for tx_skip mode in the future.

Change-Id: I372973c7518cf385f6e542b22d0f803016e693b0