vp9_ethread: the tile-based multi-threaded encoder
Currently, VP9 supports column-tile encoding, which allows a frame
to be encoded in multiple column tiles independently. The number of
column tiles are set by encoder option "--tile-columns". This
provides a way to encode a frame in parallel.

Based on previous set of patches, this patch implemented the multi-
threaded encoder.

Usage:
For HD clips:
--tile-columns=2 --threads=1/2/3/4

Change-Id: Ied987f8f2618b1283a8643ad255e88341733c9d4