Revert "Added high bitdepth sse2 transform functions"
This reverts commit 7e40a55ef974b81b2ea401702b9522c862648fe0

You need to change the pass by value parameters after the 3rd parameter to pass by const * for Visual Studio. I would change all the parameters for consistency. Here is the 64 bit doc [1] which passes 4 by register. The limit on 32 bit is 3. (but I can't find the doc that mentions it currently)

[1] http://msdn.microsoft.com/en-us/library/zthk2dkh

Change-Id: I113d73e0a62ecbb2a9deb55d18b4d598dcee9829