Proof of concept: allow intrinsics and assembly to coexist.
Function opt from https://gerrit.chromium.org/gerrit/71497

The test logic is a little convoluted and the helper file is not the
cleanest. However, this function actually does not require this
treatment (the intrinsincs are good enough to replace the assembly) and
is being used only to demonstrate the coexistence of the neon and
assembly with the preference for assembly.

Change-Id: Idde266cd7287bb6bee016c90efeafa67550f94c6