(WIP) Implement on-demond MODE_INFO allocation for decoder.
Separate the MODE_INFO management between the encoder and the decoder
as the decoder does not need to allocate MODE_INFO for all the 8x8
blocks but the encoder needs them in order to make encode decisions.

On the encoder side, a big MODE_INFO array is allocated.
On the decoder side,  a big MODE_INFO pointer array is allocated and
each 8x8 subblock inside an encoded block (64x64, 64x16…) will be pointed
to an on-demand allocated MODE_INFO. When the decoder tries to allocate a
free MODE_INFO, it first tries to fetch from free_mi_list_head. If
free_mi_list_head does not have anything available, the decoder will allocate
a new MODE_INFO and attach the new MODE_INFO to used_mi_list. When the decoder
finishes decoding a frame, it will append  all the MODE_INFO in used_mi_list
to the free_mi_list. So the next frame's decoding could reuse the allocated MODE_INFO.

This scheme will save a lot of memory on the decoder side and make the MODE_INFO
deallocation and reallocation much faster during resolution change due to MODE_INFO reuse.

Change-Id: I56e8628e9cdc67cd7e89cee66b21eca4d86fc7c5