Adjust active maxq for KF groups.
In this patch the active max Q is adjusted for each kf group
based on the vbr bit allocation and raw first pass group error.

This will tend to give a lower q for very easy sections
and a higher value for very hard sections. As such it is expected
to help improve quality in some of the easier sections.

This will almost certainly hurt overall psnr but help average.
However, most of our current reported visual problems are in
relatively easy sections which seem not to be getting enough bits.

Change-Id: Ic9befe077a602e1e7ac78103f2b214bc1fa10275