Adds an inter-intra combination mode
A smooth weighting scheme is used to put more weight
on the intra predictor samples near the left/top boundaries
and decaying it to favor the inter predictor samples more as
we move away from these boundaries in the direction of
prediction.

Results: derflr: +0.609%

WIP, since the patch does not work with the other expts yet.

Change-Id: Ic9dbe599ad6162fb05900059cbd6fc88b203a09c