Rate control adjustment.
Take into account two_pass_vbrmax_section and two_pass_vbrmin_section
parameters in vp9_vbr_rate_correction().

For example if a frame has already been allocated a number of bits >=
to the limit defined by two_pass_vbrmax_section, then we should not be
giving it more bits, even if the clip as a whole is undershooting.

Change-Id: I2575c19288dd47a68c9b2fdc7db9d1ccca6d81b1