Force splitting to 8x8 in variance partiton, for speed 6 non-rd.
During the variance partitioning (based on 8x8 averaged inputs),
8x8 block (pixel-level) variance is also computed and if found to be high
then we force the corresponding 16x16 block to be split.

This is an attempt to force smaller partitions near coding artifacts,
that can be missed by using variance based on averaged inputs.

The force splitting check is only done for superblocks that are labelled
as segment 1 (in cyclic refresh), to reduce complexity.

Change-Id: I902fec6ea2098648e9c5af88bdf55986a1d31b4c