Update patch set 1
Patch Set 1:

Just for reference - The idea is to get the same kind of benefit we get from alt ref in 2 pass in realtime. Basically I attempt to provide a strong temporal blur to the frame in hopes that it will provide a better predictor.  I make it non showable so that if we blur to strongly the subsequent frames can just choose not to use it...   

This is just play in progress.  Its Not an actual cl ready for review.     I put reviewers on as an example of what I was trying to achieve.

Patch-set: 1