Add palette coding mode for UV
A block may have two palettes, one is for Y, and the
other is for UV. Also tried buidling a signle palette
for YUV, but found it perform badly. The creation
and data structure of UV palette is very much like
Y palette.

Also fixed a bug when palette and tx-skip are both on.

on derflr
--enable-palette                    +0.00% (was -0.03%)
with all experiments                +5.94% (was +5.95%)

on screen_content
--enable-palette                    +6.02% (was +5.75%)
--enable-palette --enable-tx_skip   +15.3% (was +15.04)

Change-Id: I7287090aecc90eebcd4335d132a8c2c3895dfdd4