WIP: experiment min and max for sub 8x8 variance
This is an attempt at catching when there is high variance in a
sub 8x8 block and causing smaller block sizes.

It calculates the min and max value of an 8x8 block. If the difference
is > than a threshold we use the variance of those 2 values ( min and
max) as the 8x8 variance,   otherwise we use the single average value
as before.

Change-Id: Ie4036a24041a61405d4ceb4e6fd6b227250637ab