Update patch set 1
Patch Set 1:

It is ok to do it that way. But, since we still need to memset qcoeff/dqcoeff and eob, which is already handled in this function, it is easier to call it.

Patch-set: 1
Reviewer: Gerrit User 1000884 <1000884@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0