Non-uniform quantization experiment
The lower quantization bins are shrunk down to match closer the
source distribution (assuming a generalized gaussian-like
central peaky model for the coefficients). Specifically, the width
of the bins are increased gradually from 1-4 and from 5 onwards
all bins become the same as the nominal quantization step size.

The code works fine, experiments are in progress.

Change-Id: I88429d8cb0777021bfbb689ef69b764eafb3a1de