Non-uniform quantization experiment
This frame allows lower quantization bins to be shrunk down or
expanded to match closer the source distribution (assuming a generalized
gaussian-like central peaky model for the coefficients) in an
entropy-constrained sense. Specifically, the width of the bins 0-4 are
modified as a factor of the nominal quantization step size and from 5
onwards all bins become the same as the nominal quantization step size.
The code works fine, experiments are in progress.

Change-Id: I88429d8cb0777021bfbb689ef69b764eafb3a1de