(WIP) Improved auto frame size switching
When automatic frame sizing is enabled the codec
makes the decision as to what size to encode each
frame at.

This patch improves the process that decides when
to scale down and subsequently back up again.

When an encoded frame exceeds a defined threshold,
defined as a multiple of the target data rate for
that frame, it becomes a candidate for downscaling.

The threshold data rate multiplier is specified by
the maxsection-pct command line parameter; setting
this to 175 sets the threshold at 1.75 times the
target data rate for that frame.

Change-Id: I8fbadd29f24485d0137e8fa75deb5901dbf4f95e