Separate NEARESTMV/NEARMV from mv ref to keep their precision
Previously, NEARESTMV and NEARMV are obtained after the mv ref
candidates lower their precision, if needed. NEARESTMV is then taken
directly as the mv ref of one block, and hence NEARESTMV is always
identical to the mv ref for any block of size >= BLOCK_8X8. In this
patch, NEARESTMV and NEARMV are both obtained before the adaptive
precision checking, hence they will always keep their precision. The mv
ref of one block is still obtained after the precision checking.
NEARESTMV is hence separate from the mv ref and they do not have to be
identical to each other.

Change-Id: Idb896504a599ccf996b242b1e51178581528b8df