Update patch set 1
Patch Set 1:

The current denoiser unittest (in dataratetest) would've caught this. Problem was that cpu_speed was not set for first frame (where the crash occurs) in the unittest. It was only set from frame#=1, and so it was defaulting to speed 0 for first frame, and denoiser only works for speed >=5.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0