Update patch set 1
Patch Set 1: (1 inline comment)

The underlying bug was that the denoiser was called irrespective of frame type; it should not be called for KFs.

The bug was masked in the unit test because the KF was encoded at speed 0, which disables the denoiser. The overnight runs did not therefore pick up the bug.

Changing the unit test to set speed 5 for the KF highlighted the bug--it crashed. Marco's change should prevent the unit test failing, but we should also put code in to prevent the assert/crash if the denoiser is called on the KF.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0