aq-mode=3: Don't reset segment if block is determined to be skin.
For coding block sizes <=16X16, if the block is determined to be skin,
then always allow for that block to be candidate for refresh. So if that
block happens to be on the boost segment(s), segment won't get reset to 0
and delta-q will be applied.

Also, when setting segment map prior to encoding the frame, allow all blocks
(which were not refreshed/boosted in previous encoding frame) to be possible
candidates for refresh.

Small visual improvement, little/no change in metrics or speed.

Change-Id: I1445c27362bf2cbaa307ada88feed83c96de5956