Add skin detection in aq-mode=3.
Don't reset segment if block (for size <= 16x16) is determined to be skin.
Also, when setting segment map prior to encoding the frame, allow all blocks
(which were not refreshed/boosted in previous encoding frame) to be possible
candidates for refresh.

Small visual improvement, little/no change in metrics or speed.

Change-Id: I1445c27362bf2cbaa307ada88feed83c96de5956