tx-skip experiment: improve entropy coding of coeff tokens
This patch allows the prediction residues of tx-skipped blocks
to use probs that are different from regular transfrom
coefficients for token entropy coding. Prediction residues are
assumed as in band 6.

The initial value of probs is obtained with stats from limited
tests. The statistic model for constrained token nodes has not
been optimized. The probs for token extra bits have not been
optimized. These can be future work.

derflr with all experiments (no highbitdepth):         + 6.33%
screen_content with palette and intrabc:               +23.23%

Change-Id: I1c0d78178ee9f3655febb6f30cdaef8ee9f8e3cc