An experiment introducing a bilateral loop filter
Adds a framework to incorporate a parameterized non-linear
filter in the coding loop after the application of the
standard deblocking loop filter.

The first version uses a simplified bilateral-like filter
where the parameters conveyed are analogous to spatial and
intensity gaussian variances.

There is small gain of close of 0.1%.
WIP.

Change-Id: I20d47285b4d25b8c6386ff8af2a75ff88ac2b69b