An experiment introducing a bilateral loop filter
Adds a framework to incorporate a parameterized non-linear
filter in the coding loop after the application of the
standard deblocking loop filter.

The first version uses a straight bilateral filter
where the parameters conveyed are just spatial and
intensity gaussian variances.

Results on derflr: +0.411% (only with this experiment)
WIP - experiments with adaptive parameters.

Change-Id: I20d47285b4d25b8c6386ff8af2a75ff88ac2b69b