Expose params min-gf-interval and max-gf-interval
Adds two new vp9 parameters --min-gf-interval and --max-gf-interval
to enable testing based on frequency of alt-ref frames.

(Additional note by paulwilkins)
Note that there is a slight oddity in that key frames are also GFs and
considered part of  GF only group. However they are treated as not
being part of an arf group because for arf groups the previous GF is
assumed to be the terminal or overlay frame for the previous group.

In multi-arf mode a second arf will be placed in the middle of each
group but only if the group length is at least 6 frames.
(end note)

Change-Id: Ibf0c30b72074b3f71918ab278ccccc02a95a70a0