Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

What's the benefit to these parameters? We are just about to release and it would be nice to not immediately break the API.

If these are absolutely necessary please break the change in two - one which adds the new parameters but ignores them, and can be easily merged to the release branch, and one which implements them and has a test.

Patch-set: 3
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0