Adds a 5-tap loop postfilter
Adds a 5-tap loop postfilter that can be added optionally after the
bilateral postfilter. The filter taps are determined by standard
LMMSE estimation based on the input frame - which is the deblocked,
bilateral filtered frame, and the expected source frame. The filter
parameters are then quantized and transmitted to the decoder.

Also includes some cleanups for speed-up in the bilateral filtering
code.

WIP - not ready to be fully reviewed yet.

Change-Id: I5d37e016088e559c16317789cfb1c2f49334b2b9