Update patch set 1
Patch Set 1:

Hangyu, sorry to hijack this, but it bothered me after we went through it.
Jingning, just fyi: you weren't the source of this, but you did propagate it.

no timing yet, but I expect an improvement. does any other place this type of pattern was used come to mind?

Patch-set: 1