Update patch set 1
Patch Set 1:

> you meant |a << b| is undefined if |b < 0|, right?

|a << b| is undefined if b is less than zero, but it's also undefined if a is less than 0. 

http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf

Section 6.5.7.4:

"The result of E1 << E2 is E1 left-shifted E2 bit positions; vacated bits are filled with
zeros. If E1 has an unsigned type, the value of the result is E1 × 2^E2, reduced modulo
one more than the maximum value representable in the result type. If E1 has a signed
type and nonnegative value, and E1 × 2^E2 is representable in the result type, then that is
the resulting value; otherwise, the behavior is undefined."

Patch-set: 1
Label: Code-Review=0