A frequency and spatial domain compound coding experiment
This experiment splits the residue signal into two parts. One part
is encoded using transform coding and the other is coded
directly in pixel domain without transform.

Current method of splitting the signal is based on a two stage
scheme. In the first stage, transform is applied on the residue
signal, and the coefficients are quantized coarsely. We then use
the dequantized coefficients to obtain reconstructed signal. The
new resulting residue signal is then quantized directly without
transform as the second stage. A lot of codes are re-used from
the tx-skip experiment.

Currently no gain on derflr. On screen_content, 0.3% gain
on top of tx-skip.

To be continued. Not yet compatible with other experiments
except tx-skip.

Change-Id: Ie2e872dcc73ac8526e25cc9a2b46cba714564ba3